Starting implementation of localapp
I started but , just started nothing really done for the moment
TODO first of all : Create an entity in latest bundle (localrepo) as
localapp.
a local app is same as an app in the shop unless it is local
(obviously).

I just finish my week ! I’m not going to work on for the next 3 days,
rendez-vous prévu Monday